# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "ATP-TDM"
  },
  "security" : [
    {
      "Bearer Authentication" : [ ]
    }
  ],
...
  "paths" : {
...
    "/api/tdm/environments/{environmentId}/systems/lazy" : {
      "get" : {
        "operationId" : "getLazySystems",
        "parameters" : [
          {
            "description" : "Environment ID.",
            "in" : "path",
            "name" : "environmentId",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/LazySystem"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List the systems of an environment",
        "tags" : [
          "environments-controller"
        ]
      }
    },
...
    "/api/tdm/projects/{projectId}/environments/lazy" : {
      "get" : {
        "description" : "Returns each environment of the project with the IDs of its systems.",
        "operationId" : "getLazyEnvironments",
        "parameters" : [
          {
            "description" : "Project ID.",
            "in" : "path",
            "name" : "projectId",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/LazyEnvironment"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List the environments of a project",
        "tags" : [
          "environments-controller"
        ]
      }
    },
    "/api/tdm/projects/{projectId}/environments/lazy/refresh" : {
      "get" : {
        "description" : "Returns the same list as GET /api/tdm/projects/{projectId}/environments/lazy. Environments are read from the database on each request.",
        "operationId" : "getLazyEnvironmentsRefresh",
        "parameters" : [
          {
            "description" : "Project ID.",
            "in" : "path",
            "name" : "projectId",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/LazyEnvironment"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Reload the environments of a project",
        "tags" : [
          "environments-controller"
        ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "LazyEnvironment" : {
        "type" : "object",
        "properties" : {
          "clusterName" : {
            "type" : "string"
          },
          "created" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "modified" : {
            "type" : "string"
          },
          "modifiedBy" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "projectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "systems" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "LazySystem" : {
        "type" : "object",
        "properties" : {
          "connections" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "created" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "environmentIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "modified" : {
            "type" : "string"
          },
          "modifiedBy" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}